+++ DAMM.h
@@ -3,7 +3,7 @@
 //    FILE: DAMM.h
 //  AUTHOR: Rob Tillaart
 //    DATE: 2025-00-00
-// VERSION: 0.1.0
+// VERSION: 0.2.0
 // PURPOSE: Arduino library for calculating DAMM checksum.
 //     URL: https://github.com/RobTillaart/DAMM
 //     URL: https://en.wikipedia.org/wiki/Damm_algorithm
@@ -85,9 +85,7 @@ public:
   //  for digits as char '0'..'9'
   char add(char c)
   {
-    _count++;
-    _value = _dammMatrix[_value][c - '0'];
-    return _value + '0';
+    return add(c - '0');
   }
 
   uint32_t count()
@@ -100,8 +98,7 @@ protected:
   int       _value = 0;
   uint32_t  _count = 0;
 
-  //  PROGMEM ?
-  //  2 digits in 1 byte?
+  //  FAST MATRIX
   uint8_t _dammMatrix[10][10] =
   {
     {0, 3, 1, 7, 5, 9, 8, 6, 4, 2},
@@ -116,6 +113,113 @@ protected:
     {2, 5, 8, 1, 4, 3, 6, 7, 9, 0}
   };
 
+};
+
+
+//////////////////////////////////////////
+//
+//  ALTERNATIVE CLASS (experimental)
+//
+//  less RAM version - slower
+//
+class DAMM_small
+{
+public:
+  DAMM_small()
+  {
+    _value = 0;
+    _count = 0;
+  }
+
+  //////////////////////////////////////////////////////////////
+  //
+  //  NORMAL INTERFACE
+  //
+  bool isValid(const char * buffer)
+  {
+    return isValid((char*)buffer);
+  }
+
+  bool isValid(char * buffer)
+  {
+    _value = 0;
+    int len = strlen(buffer);
+    for (int i = 0; i < len; i++)
+    {
+      add(buffer[i] - '0');
+    }
+    return (_value == 0);
+  }
+
+  char generateChecksum(const char * buffer)
+  {
+    return generateChecksum((char *) buffer);
+  }
+
+  char generateChecksum(char * buffer)
+  {
+    _value = 0;
+    uint32_t len = strlen(buffer);
+    for (uint32_t i = 0; i < len; i++)
+    {
+      add(buffer[i] - '0');
+    }
+    return (_value + '0');
+  }
+
+
+  //////////////////////////////////////////////////////////////
+  //
+  //  STREAM INTERFACE
+  //
+  char reset()
+  {
+    int val = _value;
+    _value = 0;
+    _count = 0;
+    return val + '0';
+  }
+
+  //  for digits
+  char add(int x)
+  {
+    _count++;
+    _value = _dammMatrix_small[_value][x/2];
+    _value = (x & 1) ? _value & 0x0F : _value >> 4;
+    return _value + '0';
+  }
+
+  //  for digits as char '0'..'9'
+  char add(char c)
+  {
+    return add(c - '0');
+  }
+
+  uint32_t count()
+  {
+    return _count;
+  }
+
+
+protected:
+  int       _value = 0;
+  uint32_t  _count = 0;
+
+
+  //  SMALL MATRIX
+  uint8_t _dammMatrix_small[10][5] =
+  {
+    {0x03, 0x17, 0x59, 0x86, 0x42},
+    {0x70, 0x92, 0x15, 0x48, 0x63},
+    {0x42, 0x06, 0x87, 0x13, 0x59},
+    {0x17, 0x50, 0x98, 0x34, 0x26},
+    {0x61, 0x23, 0x04, 0x59, 0x78},
+    {0x36, 0x74, 0x20, 0x95, 0x81},
+    {0x58, 0x69, 0x72, 0x01, 0x34},
+    {0x89, 0x45, 0x36, 0x20, 0x17},
+    {0x94, 0x38, 0x61, 0x72, 0x05},
+    {0x25, 0x81, 0x43, 0x67, 0x90}
+  };
 };