+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.0] - 2025-10-28
 - initial version
+- added experimental DAMM_small class to reduce RAM.